Added ui automator test case
